Migrate Tag components's tests and stories to TS
